- Fixed OkCoin orderbook implementation - Fixed LakeBtc orderbook implementation - Fixed Bit2C orderbook implementation - Fixed BitBay orderbook implementation - Implemented GetOrderBookAsync in EthexIndiaProvider - Implemented TestGetOrderBook in EthexIndiaTests - Implemented GetOrderBookAsync in LiveCoinProvider - Implemented TestGetOrderBook in LiveCoinTests

+++ Prime.Plugins/Services/LiveCoin/LiveCoinSchema.cs
@@ -24,8 +24,8 @@ namespace Prime.Plugins.Services.LiveCoin
         internal class OrderBookResponse
         {
             public long timestamp;
-            public decimal[][] bids;
-            public decimal[][] asks;
+            public float[][] bids;
+            public float[][] asks;
         }
     }
 }
